Fix type

+++ src/host_fns/wrap_R.cu
@@ -32,7 +32,7 @@ chain_t Rchain_wrap(SEXP Rchain){
          *pi   = REAL(VECTOR_ELT(Rchain, 5)),
          *tau2 = REAL(VECTOR_ELT(Rchain, 6)),
          *zeta = INTEGER(VECTOR_ELT(Rchain, 7));
-  int *C = INTEGER(VECTOR_ELT(Rchain, 8));
+  int *C = REAL(VECTOR_ELT(Rchain, 8));
   double *probs = REAL(VECTOR_ELT(Rchain, 9)),
          *means = REAL(VECTOR_ELT(Rchain, 10)),
          *meansquares = REAL(VECTOR_ELT(Rchain, 11));
